backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
celery==5.3.1
redis==5.0.1
prometheus-client==0.18.0
pysnmp==4.4.12
netmiko==4.2.0
pandas==2.0.3
plotly==5.15.0
python-dotenv==1.0.0
django-cors-headers==4.2.0